refactor(task_6):move custom  decorator and Mixxin to  file named Perm_handler
